refactor: move webauthn devices settings section to a new component
